refactor(rebuild): Phase 0 complete — extract scrape-all, post-scrape, letterboxd-import, post-deploy-verify + scraper registry
Continues Phase 0 of the local-scraping rebuild. All four remaining
trigger task bodies are now pure-Node modules in src/lib/jobs/, and
the trigger/ wrappers are thin shims. Built in parallel by 4 sub-agents.

# New pure-Node modules

- src/lib/jobs/post-scrape.ts (166 LOC) — runPostScrapeEnrichment(payload)
  for per-cinema TMDB matching after a scrape. Pure DB + matchFilmToTMDB
  loop with backoff.

- src/lib/jobs/letterboxd-import.ts (241 LOC) — runLetterboxdImport(payload)
  for matching unmatched watchlist entries via TMDB.

- src/lib/jobs/post-deploy-verify.ts (141 LOC) — runPostDeployVerify(payload)
  for post-Vercel-deploy health checks. Imports telegram from @/lib/telegram
  directly.

- src/lib/jobs/scrape-all.ts (293 LOC) — runScrapeAll() pure orchestrator.
  Replaces Trigger.dev's batch.triggerAndWait with a runWithConcurrency
  worker pool (cap=4) that calls runScraper(buildConfig()) directly via
  the new registry. Sequential waves (chains → playwright → cheerio →
  enrichment); inlined summariseRunsSince + Telegram digest preserved.

# New scraper registry

- src/scrapers/registry.ts (374 LOC) — 27 entries (3 chains + 7 Playwright
  + 16 Cheerio + 1 enrichment). Maps task ID → buildConfig() function that
  produces SingleVenueConfig | MultiVenueConfig | ChainConfig for runScraper.
  Exports getScrapersByWave() and getScraperByTaskId() helpers.

# Trigger wrappers reduced to shims

All 4 trigger files now <20 lines, just wrap the corresponding pure-Node
function in schedules.task / task. Will be deleted wholesale when src/trigger/
is removed in a later PR.

# Intentional behaviour change

The new local scrape-all does NOT call runPostScrapeEnrichment per cinema
after each scraper (the Trigger.dev path did this via tasks.trigger as
fire-and-forget). For a once-daily local pipeline, the existing daily-sweep
(4:30am UTC) already enriches all unenriched films globally, so per-cinema
post-scrape becomes redundant. Reduces scrape-all runtime; daily-sweep
catches the same films on its next run.

# Verification

- npx tsc --noEmit clean
- npx vitest run pipeline + agents — 110/110 passing
- All 4 sub-agents independently verified before reporting back
